+++ base/src/main/java/org/aya/core/term/Term.java
@@ -19,9 +19,9 @@ import org.aya.prettier.CorePrettier;
 import org.aya.pretty.doc.Doc;
 import org.aya.ref.AnyVar;
 import org.aya.ref.LocalVar;
-import org.aya.tyck.Synthesizer;
 import org.aya.tyck.env.LocalCtx;
 import org.aya.tyck.tycker.TyckState;
+import org.aya.tyck.unify.Synthesizer;
 import org.aya.util.Arg;
 import org.aya.util.error.SourcePos;
 import org.aya.util.prettier.PrettierOptions;
