+++ base/src/main/java/org/aya/tyck/unify/DoubleChecker.java
@@ -1,10 +1,9 @@
-// Copyright (c) 2020-2022 Tesla (Yinsen) Zhang.
+// Copyright (c) 2020-2023 Tesla (Yinsen) Zhang.
 // Use of this source code is governed by the MIT license that can be found in the LICENSE.md file.
 package org.aya.tyck.unify;
 
 import org.aya.core.term.*;
 import org.aya.generic.util.NormalizeMode;
-import org.aya.tyck.Synthesizer;
 import org.aya.tyck.error.BadExprError;
 import org.aya.tyck.error.TupleError;
 import org.aya.tyck.unify.TermComparator.Sub;
